@@ -1,47 +1,40 @@
 {
-  "name": "chatv1",
+  "name": "123",
   "version": "1.0.0",
-  "description": "chat",
+  "description": "",
   "main": "App.js",
   "scripts": {
-    "builddev": "cross-env NODE_ENV=development .//node_modules//.bin//webpack-dev-server --progress --colors --history-api-fallback",
-    "buildprod": "cross-env NODE_ENV=production .//node_modules//.bin//webpack --progress",
-    "browse": ".\\node_modules\\.bin\\lite-server",
-    "test": "jest"
+    "livebuild": "cross-env NODE_ENV=development ./node_modules/.bin/webpack-dev-server --progress --colors",
+    "build": "cross-env NODE_ENV=production ./node_modules/.bin/webpack --progress"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-jest": "^21.2.0",
+    "babel-core": "^6.26.0",
     "babel-loader": "^7.1.2",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
-    "cross-env": "^5.1.5",
-    "css-loader": "^0.28.7",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "html-webpack-plugin": "^2.30.1",
-    "jest": "^21.2.1",
-    "less-loader": "^4.1.0",
-    "lite-server": "^2.4.0",
-    "node-sass": "^4.9.3",
-    "react-test-renderer": "^16.1.1",
-    "sass-loader": "^6.0.6",
-    "style-loader": "^0.19.0",
-    "webpack": "^3.8.1",
-    "webpack-config": "^7.5.0",
-    "webpack-dev-server": "^2.9.4"
+    "css-loader": "^0.28.11",
+    "style-loader": "^0.21.0",
+    "mini-css-extract-plugin": "^0.4.1",
+    "webpack-cli": "^3.1.0",
+    "webpack": "^4.16.5",
+    "webpack-dev-server": "^3.1.5",
+    "uglifyjs-webpack-plugin": "^1.3.0",
+    "sass-loader": "^7.0.1",
+    "node-sass": "^4.9.0",
+    "babel-polyfill": "^6.26.0",
+    "redux-devtools-extension": "^2.13.7",
+    "cross-env": "^5.2.0",
+    "prop-types": "^15.6.2"
   },
   "dependencies": {
-    "es6-promise": "^4.1.1",
-    "isomorphic-fetch": "^2.2.1",
-    "prop-types": "^15.6.0",
-    "react": "^16.3.1",
-    "react-dom": "^16.2.0",
-    "react-draggable": "^3.0.5",
-    "react-router-dom": "^4.2.2",
-    "render-fragment": "^0.1.1"
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
+    "redux": "^4.0.1",
+    "react-redux": "^5.0.7",
+    "redux-thunk": "^2.3.0",
+    "react-router-dom": "^4.3.1"
   }
 }
